fitur update flashdata validation

+++ application/views/kurikulum/V_kurikulum.php
@@ -3,7 +3,7 @@
         <div class="col-sm-3 sidenav">
             <h4>SIDUL</h4>
             <ul class="nav nav-pills nav-stacked">
-                <li class="active"><a href="#section1">Kurikulum</a></li>
+                <li class="active"><a href="<?php echo base_url() ?>">Kurikulum</a></li>
             </ul>
         </div>
 
@@ -13,6 +13,16 @@
                 <a href="<?php echo base_url() ?>tambah" class="btn btn-success">Tambah</a>
             </div>
             <div class="main">
+                <?php if($this->session->flashdata()): 
+                    $key = array_keys($this->session->flashdata())[0];
+                    $value = $this->session->flashdata();
+                    ?>
+                    
+                    <div class="alert alert-<?php echo $key?> alert-dismissible">
+                        <a href="#" class="close" data-dismiss="alert" aria-label="close">&times;</a>
+                        <?php print_r($value[$key]) ?>
+                    </div>
+                <?php endif ?>
                 <div class="form-group mb-20px">
                     <form action="#" method="post">
                         <div class="d-flex align-items-center">
@@ -39,7 +49,7 @@
                             <th>No.</th>
                             <th>Kode Matakuliah</th>
                             <th>Nama Matakuliah</th>
-                            <th>Jenis Perkuliahan</th>
+                            <th>Sifat Perkuliahan</th>
                             <th>Semester</th>
                             <th>Jenis Matakuliah</th>
                             <th>Bobot SKS</th>
@@ -55,15 +65,15 @@
                                 <td><?php echo $no ?></td>
                                 <td><?php echo $data['kode_matakuliah'] ?></td>
                                 <td><?php echo $data['nama_matakuliah'] ?></td>
-                                <td><?php echo $data['jenis_perkuliahan'] ?></td>
+                                <td><?php echo $data['sifat_perkuliahan'] ?></td>
                                 <td><?php echo $data['semester'] ?></td>
                                 <td><?php echo $data['jenis_matakuliah'] ?></td>
                                 <td><?php echo $data['bobot_sks'] ?></td>
                                 <td>
-                                    <a href="" class="btn btn-warning">
+                                    <a href="<?php echo base_url() ?>edit/<?php echo $data['id'] ?>" class="btn btn-warning">
                                         <i class="bi bi-pencil-fill"></i>
                                     </a>
-                                    <a href="<?php echo base_url() ?>hapus/<?php echo $data['id'] ?>" class="btn btn-danger">
+                                    <a onclick="return confirm('Apa Anda yakin untuk menghapus data ini?')" href="<?php echo base_url() ?>hapus/<?php echo $data['id'] ?>" class="btn btn-danger">
                                         <i class="bi bi-trash-fill"></i>
                                     </a>
                                 </td>
